# backend/src/main/resources/openapi/vocavista-api.yaml
openapi: 3.0.3
info:
  title: Vocavista API
  version: 0.0.1
paths:
  /api/v1/admin/users:
    get:
      tags:
        - Admin
      operationId: listAdminUsers
      summary: List users for admin management
      security:
        - GoogleOAuth: []
      responses:
        '200':
          description: Users available for admin management.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersResponse'
        '403':
          description: Admin access is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/users/{id}/status:
    put:
      tags:
        - Admin
      operationId: updateAdminUserStatus
      summary: Update a non-admin user's status
      security:
        - GoogleOAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User account identifier.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserStatusUpdateRequest'
      responses:
        '200':
          description: User status was updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access is required or target user is protected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User account was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      operationId: getCurrentUser
      summary: Get the authenticated application user
      security:
        - GoogleOAuth: []
      responses:
        '200':
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/dictionary/review:
    get:
      tags:
        - Dictionary
      operationId: getDictionaryReview
      summary: Get dictionary review items
      security:
        - GoogleOAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return. Values above 10 are capped.
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 10
        - name: includeUpcoming
          in: query
          required: false
          description: Include upcoming entries instead of only currently due entries.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Dictionary review items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictionaryReviewResponse'
        '401':
          description: Authentication is required.
  /api/v1/dictionary/review/{entryId}:
    post:
      tags:
        - Dictionary
      operationId: submitDictionaryReview
      summary: Submit a dictionary review result
      security:
        - GoogleOAuth: []
      parameters:
        - name: entryId
          in: path
          required: true
          description: Dictionary entry identifier.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DictionaryReviewSubmitRequest'
      responses:
        '200':
          description: Review result was recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictionaryReviewSubmitResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Dictionary entry was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/dictionary/videos:
    get:
      tags:
        - Dictionary
      operationId: getDictionaryVideos
      summary: Get cache manifest for dictionary pronunciation videos
      security:
        - GoogleOAuth: []
      responses:
        '200':
          description: Pronunciation videos currently used by the user's dictionary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DictionaryVideoManifestResponse'
        '401':
          description: Authentication is required.
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/media/pronunciations:
    post:
      tags:
        - Media
      operationId: createPronunciation
      summary: Queue or reuse pronunciation video generation
      security:
        - GoogleOAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PronunciationRequest'
      responses:
        '202':
          description: Pronunciation video generation was queued or reused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PronunciationResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/media/pronunciations/{id}:
    get:
      tags:
        - Media
      operationId: getPronunciation
      summary: Get pronunciation video generation status
      security:
        - GoogleOAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Pronunciation asset identifier.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pronunciation video generation status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PronunciationResponse'
        '404':
          description: Pronunciation asset was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/media/pronunciations/{id}/video:
    get:
      tags:
        - Media
      operationId: getPronunciationVideo
      summary: Stream generated pronunciation video
      security:
        - GoogleOAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Pronunciation asset identifier.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Generated pronunciation video bytes.
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pronunciation asset or generated video was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/media/pronunciations/{id}/video/small:
    get:
      tags:
        - Media
      operationId: getSmallPronunciationVideo
      summary: Stream compressed pronunciation video for normal playback
      security:
        - GoogleOAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Pronunciation asset identifier.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Compressed pronunciation video bytes.
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pronunciation asset or generated video was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/words/info:
    get:
      tags:
        - Words
      operationId: getWordInfo
      summary: Get compact German word information
      security:
        - GoogleOAuth: []
      parameters:
        - name: word
          in: query
          required: true
          description: German word or fixed expression to explain.
          schema:
            type: string
            minLength: 1
            maxLength: 80
      responses:
        '200':
          description: Word information generated for the requested word.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordInfoResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: AI provider returned an invalid or unsuccessful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: AI provider is unavailable or not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/words/suggestions:
    get:
      tags:
        - Words
      operationId: getWordSuggestions
      summary: Search cached words and generated pronunciations for autocomplete
      security:
        - GoogleOAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: Word fragment to search for.
          schema:
            type: string
            minLength: 1
            maxLength: 80
      responses:
        '200':
          description: Matching cached words and generated pronunciations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordSuggestionsResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account is not approved to use app features.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: Authenticate with Google OpenID Connect.
            email: Read the user's email address.
            profile: Read the user's display name.
  schemas:
    CurrentUserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - provider
        - status
        - admin
        - functionalAccessAllowed
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          minLength: 1
        provider:
          $ref: '#/components/schemas/OAuthProvider'
        status:
          $ref: '#/components/schemas/UserStatus'
        admin:
          type: boolean
        functionalAccessAllowed:
          type: boolean
    OAuthProvider:
      type: string
      enum:
        - google
    UserStatus:
      type: string
      enum:
        - pending
        - active
        - deactivated
    AdminUsersResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserResponse'
    AdminUserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - status
        - adminListUser
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/UserStatus'
        adminListUser:
          type: boolean
    AdminUserStatusUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/UserStatus'
    DictionaryReviewResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DictionaryReviewItem'
    DictionaryReviewItem:
      type: object
      required:
        - entryId
        - wordInfoId
        - normalizedWord
        - expectedAnswer
        - translations
        - partOfSpeech
        - dueAt
      properties:
        entryId:
          type: string
          format: uuid
        wordInfoId:
          type: string
          format: uuid
        normalizedWord:
          type: string
          minLength: 1
        expectedAnswer:
          type: string
          minLength: 1
        translations:
          $ref: '#/components/schemas/LanguageTextMap'
        partOfSpeech:
          $ref: '#/components/schemas/PartOfSpeech'
        article:
          nullable: true
          $ref: '#/components/schemas/GermanArticle'
        pronunciationAssetId:
          type: string
          format: uuid
          nullable: true
        dueAt:
          type: string
          format: date-time
    DictionaryReviewSubmitRequest:
      type: object
      required:
        - correct
      properties:
        correct:
          type: boolean
    DictionaryReviewSubmitResponse:
      type: object
      required:
        - entryId
        - correct
        - expectedAnswer
        - dueAt
        - intervalDays
        - correctStreak
        - repetitionCount
        - lapseCount
      properties:
        entryId:
          type: string
          format: uuid
        correct:
          type: boolean
        expectedAnswer:
          type: string
          minLength: 1
        dueAt:
          type: string
          format: date-time
        intervalDays:
          type: integer
          minimum: 0
        correctStreak:
          type: integer
          minimum: 0
        repetitionCount:
          type: integer
          minimum: 0
        lapseCount:
          type: integer
          minimum: 0
    DictionaryVideoManifestResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DictionaryVideoManifestItem'
    DictionaryVideoManifestItem:
      type: object
      required:
        - pronunciationAssetId
        - wordInfoId
        - videoUrl
        - fullVideoUrl
        - updatedAt
      properties:
        pronunciationAssetId:
          type: string
          format: uuid
        wordInfoId:
          type: string
          format: uuid
        videoUrl:
          type: string
          format: uri
        fullVideoUrl:
          type: string
          format: uri
        updatedAt:
          type: string
          format: date-time
    PronunciationRequest:
      type: object
      required:
        - wordInfoId
        - word
        - phrase
        - language
      properties:
        wordInfoId:
          type: string
          format: uuid
        word:
          type: string
          minLength: 1
          maxLength: 80
        phrase:
          type: string
          minLength: 1
          maxLength: 240
        language:
          type: string
          enum:
            - de
    PronunciationResponse:
      type: object
      required:
        - id
        - wordInfoId
        - status
      properties:
        id:
          type: string
          format: uuid
        wordInfoId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PronunciationStatus'
        videoUrl:
          type: string
          format: uri
          nullable: true
        fullVideoUrl:
          type: string
          format: uri
          nullable: true
        errorCode:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
    PronunciationStatus:
      type: string
      enum:
        - queued
        - processing
        - completed
        - failed
    WordInfoResponse:
      type: object
      required:
        - normalizedWord
        - id
        - language
        - translations
        - partOfSpeech
        - frequency
        - isCompound
        - compoundParts
        - shortNote
        - examples
      properties:
        id:
          type: string
          format: uuid
        normalizedWord:
          type: string
          minLength: 1
        language:
          type: string
          enum:
            - de
        translations:
          $ref: '#/components/schemas/LanguageTextMap'
        partOfSpeech:
          $ref: '#/components/schemas/PartOfSpeech'
        gender:
          nullable: true
          $ref: '#/components/schemas/Gender'
        article:
          nullable: true
          $ref: '#/components/schemas/GermanArticle'
        plural:
          type: string
          nullable: true
        frequency:
          $ref: '#/components/schemas/WordFrequency'
        isCompound:
          type: boolean
        compoundParts:
          type: array
          items:
            $ref: '#/components/schemas/CompoundPart'
        shortNote:
          $ref: '#/components/schemas/LanguageTextMap'
        examples:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: '#/components/schemas/WordExample'
    WordSuggestionsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WordSuggestion'
    WordSuggestion:
      type: object
      required:
        - word
        - source
      properties:
        word:
          type: string
          minLength: 1
        phrase:
          type: string
          nullable: true
        source:
          type: string
          enum:
            - word_info
            - pronunciation
        wordInfoId:
          type: string
          format: uuid
          nullable: true
        pronunciationId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/PronunciationStatus'
          nullable: true
        videoUrl:
          type: string
          format: uri
          nullable: true
        fullVideoUrl:
          type: string
          format: uri
          nullable: true
    CompoundPart:
      type: object
      required:
        - word
        - meanings
      properties:
        word:
          type: string
          minLength: 1
        meanings:
          $ref: '#/components/schemas/LanguageTextMap'
    WordExample:
      type: object
      required:
        - sentence
        - translations
      properties:
        sentence:
          type: string
          minLength: 1
        translations:
          $ref: '#/components/schemas/LanguageTextMap'
    LanguageTextMap:
      type: object
      minProperties: 1
      additionalProperties:
        type: array
        minItems: 1
        items:
          type: string
          minLength: 1
    PartOfSpeech:
      type: string
      enum:
        - noun
        - verb
        - adjective
        - adverb
        - pronoun
        - preposition
        - conjunction
        - interjection
        - phrase
        - other
    WordFrequency:
      type: string
      enum:
        - rare
        - uncommon
        - common
        - very_common
    Gender:
      type: string
      enum:
        - masculine
        - feminine
        - neuter
    GermanArticle:
      type: string
      enum:
        - der
        - die
        - das
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string